修改FileUtils.GetSizeDisplayName的参数类型到long

+++ ZKWeb.Utils/Functions/FileUtils.cs
@@ -13,9 +13,9 @@ namespace ZKWeb.Utils.Functions {
 		/// 获取字节大小的显示名称
 		/// 自动转换成KB、MB或GB格式
 		/// </summary>
-		/// <param name="bytes"></param>
+		/// <param name="bytes">字节数</param>
 		/// <returns></returns>
-		public static string GetSizeDisplayName(int bytes) {
+		public static string GetSizeDisplayName(long bytes) {
 			if (bytes >= 1073741824) {
 				return $"{Decimal.Round((decimal)bytes / 1073741824, 2)}GB";
 			} else if (bytes >= 1048576) {
